Extracted server.rb file and refactor
