+++ app/views/listings/all.erb
